feat(minimal-api): remove Aspire.MinimalApi project and related files
