+++ src/AspireApp.MinimalApi/Endpoints/ProductEndpoints.cs
